consolidate constants and errors, move data to lib
