+++ src/lib/utils.ts
@@ -34,3 +34,35 @@ export const ErrorMessage = {
   NOTIFICATION_QUEUE_FAILED: 'Failed to queue notification',
   NOTIFICATION_PREFERENCE_UPDATE_FAILED: 'Failed to update notification preferences',
 } as const;
+
+export class Utils extends Error {
+  constructor(
+    message: string = ErrorMessage.SERVER_ERROR,
+    public statusCode: number = ErrorCode.SERVER_ERROR,
+  ) {
+    super(message);
+    this.name = this.constructor.name;
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+export class AuthenticationError extends Utils {
+  constructor(message: string = ErrorMessage.AUTHENTICATION_REQUIRED) {
+    super(message, ErrorCode.AUTHENTICATION);
+  }
+}
+
+export class AuthorizationError extends Utils {
+  constructor(message: string = ErrorMessage.INSUFFICIENT_PERMISSIONS) {
+    super(message, ErrorCode.AUTHORIZATION);
+  }
+}
+
+export class ValidationError extends Utils {
+  constructor(
+    message: string = ErrorMessage.VALIDATION_FAILED,
+    public errors?: unknown,
+  ) {
+    super(message, ErrorCode.VALIDATION);
+  }
+}
